Added support for layer visibility and cel opacity
Turning off the visibility of layers in Pixelorama will now result in them being skipped in the plugin.

Adjusting the cel's opacity in Pixelorama will also result in the cels being blended into the frame with a different opacity. Right now it has the old blending code if the opacity is equal to 1 as the opacity code should be slower.

I can create screenshots if needed.