Added support for layer visibility and cel opacity (#4)
* Added support for layer visibility and cel opacity

Turning off the visibility of layers in Pixelorama will now result in them being skipped in the plugin.

Adjusting the cel's opacity in Pixelorama will also result in the cels being blended into the frame with a different opacity. Right now it has the old blending code if the opacity is equal to 1 as the opacity code should be slower.

I can create screenshots if needed.

* Simplify opacity and optimize

Simplified the opacity by just multiplying it into the cel's 

Optimized by skipping fully transparent cels.

Optimized by using file.seek() to skip over the cel data when that cel isn't visible (before it created the image for that cel and just skipped blending it in).